勉强完成了投资组合的多因子

+++ src/main/java/com/example/ffbackend/entity/StockInPortfolio.java
@@ -20,7 +20,7 @@ public class StockInPortfolio {
 
     Integer userId;
     String code;
-    Double num;
+    Integer num;
     Double money;
 
     public StockInPortfolioVo getVo () {
